requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
Flask-WTF
Authlib
python-dotenv
pymysql
requests
Flask-Caching
